charity_system/urls.py
"""
URL Configuration for Charity Management System
"""
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenBlacklistView,
)

urlpatterns = [
    # Django Admin
    path('admin/', admin.site.urls),
    
    # HTML Frontend URLs
    path('', include('children.urls_html')),
    
    # API endpoints
    path('api/v1/', include('children.urls')),
    
    # JWT Authentication
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/token/blacklist/', TokenBlacklistView.as_view(), name='token_blacklist'),
]

# Serve media files in development
if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)

# Custom admin site header
admin.site.site_header = 'Charity Management System'
admin.site.site_title = 'Charity CMS'
admin.site.index_title = 'Children & Households Management'
